aws-iot-softwarepackageversion/src/main/java/software/amazon/iot/softwarepackageversion/ListHandler.java
package software.amazon.iot.softwarepackageversion;

import software.amazon.awssdk.services.iot.IotClient;
import software.amazon.awssdk.services.iot.model.IotException;
import software.amazon.awssdk.services.iot.model.ListPackageVersionsRequest;
import software.amazon.awssdk.services.iot.model.ListPackageVersionsResponse;
import software.amazon.awssdk.services.iot.model.ListPackagesRequest;
import software.amazon.awssdk.services.iot.model.ListPackagesResponse;
import software.amazon.awssdk.services.iot.model.PackageSummary;
import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
import software.amazon.cloudformation.proxy.Logger;
import software.amazon.cloudformation.proxy.OperationStatus;
import software.amazon.cloudformation.proxy.ProgressEvent;
import software.amazon.cloudformation.proxy.ProxyClient;
import software.amazon.cloudformation.proxy.ResourceHandlerRequest;

/**
 * API Calls for ListHandler:
 * ListPackageVersions: To list all PackageVersions in an account
 */
public class ListHandler extends BaseHandlerStd {

    private static final String OPERATION = "ListSoftwarePackageVersions";

    @Override
    protected ProgressEvent<ResourceModel, CallbackContext> handleRequest(
            final AmazonWebServicesClientProxy proxy,
            final ResourceHandlerRequest<ResourceModel> request,
            final CallbackContext callbackContext,
            final ProxyClient<IotClient> proxyClient,
            final Logger logger) {
        final ResourceModel resourceModel = request.getDesiredResourceState();
        String packageName = resourceModel.getPackageName();

        if (packageName == null || packageName.isEmpty()) {
            final ListPackagesRequest listPackagesRequest = ListPackagesRequest.builder()
                    .build();

            ListPackagesResponse listPackagesResponse =
                    proxy.injectCredentialsAndInvokeV2(listPackagesRequest, proxyClient.client()::listPackages);

            // In order to pass the contract tests, enforce one package per account
            packageName = listPackagesResponse
                    .packageSummaries().get(0).packageName();
        }

        try {
            final ListPackageVersionsRequest listPackageVersionsRequest = Translator.translateToListRequest(packageName, request.getNextToken());
            ListPackageVersionsResponse listPackageVersionsResponse = proxy.injectCredentialsAndInvokeV2(
                    listPackageVersionsRequest,
                    proxyClient.client()::listPackageVersions
            );
            String nextToken = listPackageVersionsResponse.nextToken();
            return ProgressEvent.<ResourceModel, CallbackContext>builder()
                    .resourceModels(Translator.translateFromListResponse(listPackageVersionsResponse))
                    .nextToken(nextToken)
                    .status(OperationStatus.SUCCESS)
                    .build();
        } catch (IotException e) {
            throw Translator.translateIotExceptionToHandlerException(null, OPERATION, e);
        }
    }
}
